Refactor layout and components
This commit refactors the layout component by importing the global CSS file from the app directory. It also removes the navbar and background image from the layout component. Additionally, the commit deletes the page component and updates the CompetenciasAnteriores and Footer components with some styling changes. Finally, the Noticias component is updated with new titles and URLs for the news items.